fix: Move parse_semantic_check_response to avoid kimina_client import on Python 3.11
Move parse_semantic_check_response from kimina_server.py to common.py to
fix Python 3.11 CI failures. The function is a pure string parser with
no kimina_client dependency, but importing it from kimina_server.py
triggered a PydanticUserError due to kimina_client's TypedDict
compatibility issues with Python < 3.12.

Changes:
- Move parse_semantic_check_response from kimina_server.py to common.py
- Update informal_theorem_semantics_agent.py to import from common.py
- Remove unused imports (re, LLMParsingError) from kimina_server.py

This eliminates the kimina_client import dependency for the semantics
agent, allowing tests to run on Python 3.11 without import errors.

Fixes Python 3.11 CI test failures in tests-and-type-check workflow.
